Revert "Feature/hide save when not loaded"
